+++ src/main/java/ju/controller/FindController.java
@@ -238,67 +238,67 @@ public class FindController {
 		return mav;
 	}
 	
-	@RequestMapping(value="/test.ju")
-	public ModelAndView test(){
-		File f = new File("../LOG/book/searchList.log");
-		
-		System.out.println("경로 :" + f.getAbsoluteFile().getPath());
-		Map<String, Integer> map = new HashMap<String, Integer>();
-		ValueComparator bvc =  new ValueComparator(map);
-        TreeMap<String,Integer> sorted_map = new TreeMap<String,Integer>(bvc);
-        
-		if(f.exists()){
-			try {
-				System.out.println("파일있음");
-				
-				////////////////////////////////////////////////////////////////
-				BufferedReader in = new BufferedReader(new FileReader(f));
-				String s;
-				
-				while ((s = in.readLine()) != null) {
-					System.out.println("s : "+s);
-					String arr[] = s.split("/");
-					
-					for (int i = 0; i < arr.length; i++) {
-						if(map.get(arr[i])==null){
-							map.put(arr[i], 1);
-						}else{
-							int count = map.get(arr[i])+1;
-							map.put(arr[i], count);
-							
-						}
-					}
-				}
-				in.close();
-			} catch (Exception e) {
-			}
-		}else{
-			System.out.println("파일없음");
-		}
-		
-		sorted_map.putAll(map);
-        
-		String json = "[";
-		int count = 0;
-        for (Map.Entry<String,Integer> entry : sorted_map.entrySet()) {
-            //정렬한 리스트에서 순번을 배열번호로 변경하여 원본 리스트에서 추출
-        	count++;
-        	if(count>7){
-        		break;
-    		}else{
-    			json += "{text:'" + entry.getKey() + "', count:'" + map.get(entry.getKey()) + "'},";
-    		}
-        }
-        json = json.substring(0, json.length()-1);
-        json += "]";
-		
-		
-		
-		ModelAndView mav = new ModelAndView("/find/test");
-		System.out.println(json);
-		mav.addObject("list", json);
-		return mav;
-	}
+//	@RequestMapping(value="/test.ju")
+//	public ModelAndView test(){
+//		File f = new File("../LOG/book/searchList.log");
+//		
+//		System.out.println("경로 :" + f.getAbsoluteFile().getPath());
+//		Map<String, Integer> map = new HashMap<String, Integer>();
+//		ValueComparator bvc =  new ValueComparator(map);
+//        TreeMap<String,Integer> sorted_map = new TreeMap<String,Integer>(bvc);
+//        
+//		if(f.exists()){
+//			try {
+//				System.out.println("파일있음");
+//				
+//				////////////////////////////////////////////////////////////////
+//				BufferedReader in = new BufferedReader(new FileReader(f));
+//				String s;
+//				
+//				while ((s = in.readLine()) != null) {
+//					System.out.println("s : "+s);
+//					String arr[] = s.split("/");
+//					
+//					for (int i = 0; i < arr.length; i++) {
+//						if(map.get(arr[i])==null){
+//							map.put(arr[i], 1);
+//						}else{
+//							int count = map.get(arr[i])+1;
+//							map.put(arr[i], count);
+//							
+//						}
+//					}
+//				}
+//				in.close();
+//			} catch (Exception e) {
+//			}
+//		}else{
+//			System.out.println("파일없음");
+//		}
+//		
+//		sorted_map.putAll(map);
+//        
+//		String json = "[";
+//		int count = 0;
+//        for (Map.Entry<String,Integer> entry : sorted_map.entrySet()) {
+//            //정렬한 리스트에서 순번을 배열번호로 변경하여 원본 리스트에서 추출
+//        	count++;
+//        	if(count>7){
+//        		break;
+//    		}else{
+//    			json += "{text:'" + entry.getKey() + "', count:'" + map.get(entry.getKey()) + "'},";
+//    		}
+//        }
+//        json = json.substring(0, json.length()-1);
+//        json += "]";
+//		
+//		
+//		
+//		ModelAndView mav = new ModelAndView("/find/test");
+//		System.out.println(json);
+//		mav.addObject("list", json);
+//		return mav;
+//	}
 	
 	@RequestMapping(value="/dash.ju")
 	public ModelAndView dash(){
