+++ src/main/java/ju/controller/MemberController.java
@@ -6,6 +6,7 @@ import java.io.IOException;
 import java.util.*;
 
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -14,8 +15,6 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.servlet.ModelAndView;
 
 //import ju.dto.*;
-import ju.member.model.*;
-import ju.member.model.MemberDAO;
 import ju.model.*;
 import ju.dto.*;
 
@@ -153,18 +152,26 @@ public class MemberController {
 		
 	}
 	@RequestMapping("/memberLoginOk.ju")
-	public String loginOk(
+	public ModelAndView loginOk(
 			@RequestParam(value="mem_id",defaultValue="")String mem_id,
-			@RequestParam(value="mem_pwd",defaultValue="")String mem_pwd){
+			@RequestParam(value="mem_pwd",defaultValue="")String mem_pwd,
+			HttpSession session){
+		ModelAndView mav = new ModelAndView();
+		MemberDTO dto = memberDao.loginSubmit(mem_id, mem_pwd);
 		
-		int result = memberDao.loginSubmit(mem_id, mem_pwd);
-		System.out.println("result : "+result);
-		if(result==1){
+		
+		if(dto.getMem_name().equals("nolog")||dto.getMem_name().equals("black")){
 			
-			return "index";
+			mav.setViewName("member/memberLogin");
+			return mav;
+			
 		}else{
+			session.setAttribute("sid", dto.getMem_id());
+			session.setAttribute("sname", dto.getMem_name());
+			session.setAttribute("sidx", dto.getMem_idx());
+			mav.setViewName("index");
+			return mav;
 			
-			return "member/memberLogin";
 		}
 		
 	}
