+++ src/main/java/ju/controller/adminBookController.java
@@ -13,7 +13,6 @@ import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
-import javax.servlet.http.HttpSession;
 
 import java.io.*;
 import java.net.*;
@@ -28,7 +27,7 @@ public class adminBookController {
 	public LoanDAO loanDao;
 	
 	@Autowired
-	public ju.member.model.MemberDAO memberDao;
+	MemberDAO memberDao;
 	
 	@Autowired
 	public YeyakDAO yeyakDao;
