+++ src/main/java/ju/controller/adminFedexController.java
@@ -16,7 +16,7 @@ public class adminFedexController {
 	public FedexDAO fedexDao;
 	
 	@Autowired
-	public ju.member.model.MemberDAO memberDao;
+	MemberDAO memberDao;
 	
 	@Autowired
 	public BookDAO bookDao;
