+++ src/main/java/ju/controller/adminLoanBookController.java
@@ -36,7 +36,7 @@ public class adminLoanBookController {
 	public FedexDAO fedexDao;
 	
 	@Autowired
-	public ju.member.model.MemberDAO memberDao;
+	MemberDAO memberDao;
 	
 	
 	// 대출관리 메인페이지로 이동
