+++ src/main/webapp/WEB-INF/dispatcher-servlet.xml
@@ -32,7 +32,7 @@
 		<constructor-arg name="sqlMap" ref="sqlSessionTemplate"></constructor-arg>
 	</bean>
 	<!-- service -->
-	<bean id="memberDao" class="ju.member.model.MemberDAOImple">
+	<bean id="memberDao" class="ju.model.MemberDAOImple">
 		<constructor-arg ref="sqlSessionTemplate" />
 	</bean>
 	<bean id="bigdao" class="ju.booking.model.BookingDAOImple">
