+++ src/main/webapp/WEB-INF/views/header.jsp
@@ -151,6 +151,7 @@
 						<li><a href="/lee/libMap.ju">오시는길</a></li>
 						<li><a href="/lee/libStatus.ju">장서정보(소장도서정보)</a></li>
 						<li><a href="/lee/libGuide.ju">이용안내</a></li>
+						<li><a href="/lee/analysis.ju">도서관 통계</a></li>
 					</ul>
 				</li>
 				<li><a href="/lee/bkList.ju">일반도서 검색</a></li>
